pyproject.toml
[project]
name = "semantic-model-weaver"
version = "0.1.0"
description = "Agentic Cortex Analyst semantic model generator and evaluator"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.2",
    "pyyaml>=6.0.3",
    "rich>=13.9.4",
    "snowflake-snowpark-python",
    "streamlit>=1.56.0",
    "trulens-connectors-snowflake",
    "trulens-core",
    "trulens-feedback",
    "trulens-providers-cortex",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-mock>=3.15.1",
    "requests>=2.33.1",
    "ruff>=0.15.9",
    # snowflake-connector-python is brought in transitively by trulens packages
]

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "integration: requires live Snowflake connection (SNOWFLAKE_PASSWORD / SNOWSQL_PWD must be set)",
]
log_cli = true
log_cli_level = "INFO"

